Replace galley model with updated geometry
