+++ src/main/java/com/talhanation/smallships/client/model/ModelGalley.java
@@ -1,6 +1,5 @@
 package com.talhanation.smallships.client.model;
 
-import com.google.common.collect.ImmutableList;
 import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.vertex.IVertexBuilder;
 import com.talhanation.smallships.entities.AbstractBannerUser;
@@ -8,1211 +7,421 @@ import com.talhanation.smallships.entities.GalleyEntity;
 import net.minecraft.client.renderer.entity.model.EntityModel;
 import net.minecraft.client.renderer.model.ModelRenderer;
 import net.minecraft.util.math.MathHelper;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 
-@OnlyIn(Dist.CLIENT)
 public class ModelGalley extends EntityModel<AbstractBannerUser> {
-   public ModelRenderer botom_1;
-   public ModelRenderer side_1;
-   public ModelRenderer side_3;
-   public ModelRenderer botom_2;
-   public ModelRenderer botom_3;
-   public ModelRenderer botom_4;
-   public ModelRenderer side_4;
-   public ModelRenderer side_5;
-   public ModelRenderer side_6;
-   public ModelRenderer side_7;
-   public ModelRenderer side_8;
-   public ModelRenderer side_9;
-   public ModelRenderer side_10;
-   public ModelRenderer side_2;
-   public ModelRenderer botom_1_1;
-   public ModelRenderer botom_5;
-   public ModelRenderer botom_7;
-   public ModelRenderer mast_1;
-   public ModelRenderer seil_1_1;
-   public ModelRenderer seil_2_1;
-   public ModelRenderer side_11;
-   public ModelRenderer side_12;
-   public ModelRenderer Cargo0;
-   public ModelRenderer Cargo1;
-   public ModelRenderer ruder_l_3;
-   public ModelRenderer ruder_l_1;
-   public ModelRenderer side_18;
-   public ModelRenderer side_19;
-   public ModelRenderer ruder_l_2;
-   public ModelRenderer ruder_l_4;
-   public ModelRenderer botom_6;
-   public ModelRenderer botom_8;
-   public ModelRenderer botom_9;
-   public ModelRenderer side_13;
-   public ModelRenderer ruder_r_1;
-   public ModelRenderer ruder_r_2;
-   public ModelRenderer ruder_r_3;
-   public ModelRenderer ruder_r_4;
-   public ModelRenderer Mast_unten;
-   public ModelRenderer Mast_oben;
-   public ModelRenderer mast_2;
-   public ModelRenderer seil_4_1;
-   public ModelRenderer mast_3;
-   public ModelRenderer mast_3_1;
-   public ModelRenderer mast_4;
-   public ModelRenderer seil_4_2;
-   public ModelRenderer seil_4_3;
-   public ModelRenderer seil_4_4;
-   public ModelRenderer seil_4_4_2;
-   public ModelRenderer seil_4_4_3;
-   public ModelRenderer seil_1_2;
-   public ModelRenderer seil_1_3;
-   public ModelRenderer seil_1_4;
-   public ModelRenderer seil_1_4_1;
-   public ModelRenderer seil_1_4_2;
-   public ModelRenderer seil_1_4_3;
-   public ModelRenderer seil_1_4_4;
-   public ModelRenderer seil_1_4_5;
-   public ModelRenderer seil_2_2;
-   public ModelRenderer seil_2_3;
-   public ModelRenderer seil_2_4;
-   public ModelRenderer seil_2_4_1;
-   public ModelRenderer seil_2_4_2;
-   public ModelRenderer seil_2_4_3;
-   public ModelRenderer Cargo_0_1;
-   public ModelRenderer Cargo_0_2;
-   public ModelRenderer Cargo_1_1;
-   public ModelRenderer Cargo_1_2;
-   public ModelRenderer Segel_z0;
-   public ModelRenderer Segel_z4;
-   public ModelRenderer Segel_z3;
-   public ModelRenderer Segel_z2;
-   public ModelRenderer Segel_z1;
-   public ModelRenderer Segel_1_2;
-   public ModelRenderer Segel_1_3;
-   public ModelRenderer Segel_1_4;
-   public ModelRenderer Segel_1_5;
-   public ModelRenderer Segel_1_6;
-   public ModelRenderer Segel_1_7;
-   public ModelRenderer Segel_1_8;
-   public ModelRenderer Segel_1_9;
-   public ModelRenderer Segel_1_10;
-   public ModelRenderer Segel_1_11;
-   public ModelRenderer Segel_1_12;
-   public ModelRenderer Segel_1_13;
-   public ModelRenderer Segel_1_14;
-   public ModelRenderer Segel_1_15;
-   public ModelRenderer Segel_1_16;
-   public ModelRenderer Segel_1_17;
-   public ModelRenderer Segel_1_18;
-   public ModelRenderer Segel_1_2_1;
-   public ModelRenderer Segel_1_3_1;
-   public ModelRenderer Segel_1_4_1;
-   public ModelRenderer Segel_1_5_1;
-   public ModelRenderer Segel_1_6_1;
-   public ModelRenderer Segel_1_7_1;
-   public ModelRenderer Segel_1_8_1;
-   public ModelRenderer Segel_1_9_1;
-   public ModelRenderer Segel_1_10_1;
-   public ModelRenderer Segel_1_11_1;
-   public ModelRenderer Segel_1_12_1;
-   public ModelRenderer Segel_1_13_1;
-   public ModelRenderer Segel_1_2_2;
-   public ModelRenderer Segel_1_3_2;
-   public ModelRenderer Segel_1_4_2;
-   public ModelRenderer Segel_1_5_2;
-   public ModelRenderer Segel_1_6_2;
-   public ModelRenderer Segel_1_7_2;
-   public ModelRenderer Segel_1_8_2;
-   public ModelRenderer Segel_1_13_2;
-   public ModelRenderer Segel_1_2_3;
-   public ModelRenderer Segel_1_3_3;
-   public ModelRenderer Segel_1_4_3;
-   public ModelRenderer Segel_1_5_3;
-   public ModelRenderer Segel_1_13_3;
+    private static final float HULL_Y_ROT = ((float) Math.PI / 2F);
+    private static final float PADDLE_MIN_X_ROT = -1.0471976F;
+    private static final float PADDLE_MAX_X_ROT = -0.2617994F;
+    private static final float PADDLE_MIN_Y_ROT = -0.7853982F;
+    private static final float PADDLE_MAX_Y_ROT = 0.7853982F;
 
-   private static final float HULL_Y_ROT = ((float)Math.PI / 2F);
-   private static final float PADDLE_MIN_X_ROT = -1.0471975803375244F;
-   private static final float PADDLE_MAX_X_ROT = -0.2617993950843811F;
-   private static final float PADDLE_MIN_Y_ROT = -0.7853981852531433F;
-   private static final float PADDLE_MAX_Y_ROT = 0.7853981852531433F;
+    private final ModelRenderer root;
+    private final ModelRenderer galley;
+    private final ModelRenderer steer;
+    private final ModelRenderer[] chests;
+    private final ModelRenderer[] leftPaddles;
+    private final ModelRenderer[] rightPaddles;
 
-   private final ModelRenderer[] leftPaddles;
-   private final ModelRenderer[] rightPaddles;
-   private final ModelRenderer[] sailStates;
+    public ModelGalley() {
+        this.texWidth = 128;
+        this.texHeight = 64;
 
-   public ModelGalley() {
-      this.texWidth = 128;
-      this.texHeight = 64;
-      this.side_6 = new ModelRenderer(this, 0, 27);
-      this.side_6.setPos(61.0F, 4.0F, 16.0F);
-      this.side_6.addBox(-8.0F, -7.0F, -1.0F, 16.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_6, 0.0F, 1.5707964F, 0.0F);
-      this.side_7 = new ModelRenderer(this, 0, 27);
-      this.side_7.setPos(61.0F, 4.0F, 0.0F);
-      this.side_7.addBox(-8.0F, -7.0F, -1.0F, 16.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_7, 0.0F, 1.5707964F, 0.0F);
-      this.seil_1_3 = new ModelRenderer(this, 0, 0);
-      this.seil_1_3.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_1_3.texOffs(100, 0).addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_7_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_7_1, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_1_8 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_8.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_8, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.seil_2_1 = new ModelRenderer(this, 96, 0);
-      this.seil_2_1.setPos(73.5F, -3.0F, 7.5F);
-      this.seil_2_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.seil_2_1, 1.5707964F, 0.0F, -0.6806784F);
-      this.Cargo_1_1 = new ModelRenderer(this, 30, 55);
-      this.Cargo_1_1.setPos(2.0F, -5.0F, 1.0F);
-      this.Cargo_1_1.addBox(0.0F, 0.0F, 0.0F, 4.0F, 5.0F, 4.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5 = new ModelRenderer(this, 100, 0);
-      this.Segel_1_5.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_5, -0.033161256F, 0.010471976F, 0.08726646F);
-      this.Cargo0 = new ModelRenderer(this, 96, 38);
-      this.Cargo0.setPos(-13.0F, -5.0F, -5.0F);
-      this.Cargo0.addBox(0.0F, 0.0F, -2.0F, 8.0F, 8.0F, 8.0F, 0.0F, 0.0F, 0.0F);
-      this.mast_4 = new ModelRenderer(this, 10, 0);
-      this.mast_4.setPos(0.0F, 74.0F, 0.0F);
-      this.mast_4.addBox(0.0F, 0.0F, 0.0F, 3.0F, 21.0F, 4.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_4_2.setPos(0.0F, 0.0F, -2.0F);
-      this.Segel_1_4_2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_4_2, -0.033161256F, 0.010471976F, 0.13962634F);
-      this.mast_2 = new ModelRenderer(this, 10, 0);
-      this.mast_2.setPos(0.0F, -18.0F, 0.0F);
-      this.mast_2.addBox(0.0F, 0.0F, 0.0F, 3.0F, 18.0F, 4.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1 = new ModelRenderer(this, 100, 0);
-      this.Segel_1_5_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_5_1, -0.033161256F, 0.010471976F, 0.08726646F);
-      this.Segel_1_8_2 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_8_2.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_8_2, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_1_18 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_18.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_18.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_18, -0.041887905F, 0.010471976F, 0.17453292F);
-      this.side_8 = new ModelRenderer(this, 0, 27);
-      this.side_8.setPos(-15.0F, 4.0F, 0.0F);
-      this.side_8.addBox(-8.0F, -7.0F, -1.0F, 16.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_8, 0.0F, 1.5707964F, 0.0F);
-      this.botom_9 = new ModelRenderer(this, 10, 0);
-      this.botom_9.setPos(0.0F, 3.0F, 17.0F);
-      this.botom_9.addBox(-14.0F, -9.0F, -3.0F, 18.0F, 16.0F, 3.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.botom_9, 1.5707964F, 0.0F, 0.0F);
-      this.seil_2_2 = new ModelRenderer(this, 96, 0);
-      this.seil_2_2.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_2_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.seil_1_4_2 = new ModelRenderer(this, 96, 0);
-      this.seil_1_4_2.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_1_4_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_4.setPos(0.0F, 0.0F, -2.0F);
-      this.Segel_1_4.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_4, -0.033161256F, 0.010471976F, 0.13962634F);
-      this.Segel_1_6_1 = new ModelRenderer(this, 98, 0);
-      this.Segel_1_6_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_6_1, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.seil_4_4_2 = new ModelRenderer(this, 96, 0);
-      this.seil_4_4_2.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_4_4_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_l_2 = new ModelRenderer(this, 62, 0);
-      this.ruder_l_2.setPos(17.0F, -4.0F, 25.0F);
-      this.ruder_l_2.addBox(-1.0F, 0.0F, -8.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_l_2.addBox(-0.8F, -3.0F, 9.0F, 1.0F, 6.0F, 7.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.ruder_l_2, -0.65449846F, 0.66095614F, 0.19634955F);
-      this.ruder_r_1 = new ModelRenderer(this, 62, 20);
-      this.ruder_r_1.setPos(0.0F, -4.0F, -10.0F);
-      this.ruder_r_1.addBox(-1.0F, 0.0F, -8.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_r_1.addBox(-0.2F, -3.0F, 9.0F, 1.0F, 6.0F, 7.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.ruder_r_1, -0.65449846F, 2.4806366F, 0.19634955F);
-      this.Mast_oben = new ModelRenderer(this, 0, 5);
-      this.Mast_oben.setPos(21.0F, -64.0F, 8.5F);
-      this.Mast_oben.addBox(0.0F, 0.0F, -9.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_oben.addBox(0.0F, 0.0F, -27.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_oben.addBox(0.0F, 0.0F, 9.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_oben.addBox(0.0F, 0.0F, -45.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_oben.addBox(0.0F, 0.0F, -63.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Mast_oben, 0.61086524F, -1.0471976F, 0.0F);
-      this.Segel_1_13 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_13.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_13, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.seil_2_4_1 = new ModelRenderer(this, 96, 0);
-      this.seil_2_4_1.setPos(0.0F, 0.0F, 13.0F);
-      this.seil_2_4_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.botom_6 = new ModelRenderer(this, 0, 0);
-      this.botom_6.setPos(57.0F, 6.0F, 9.0F);
-      this.botom_6.addBox(-14.0F, -9.0F, -3.0F, 16.0F, 16.0F, 3.0F, 0.0F, 4.0F, 0.0F);
-      this.setRotateAngle(this.botom_6, 1.5707964F, 0.0F, 0.0F);
-      this.side_12 = new ModelRenderer(this, 0, 27);
-      this.side_12.setPos(-19.0F, 0.0F, 8.0F);
-      this.side_12.addBox(-8.0F, -7.0F, -1.0F, 16.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_12, 0.0F, 1.5707964F, 0.0F);
-      this.seil_1_4_5 = new ModelRenderer(this, 110, 0);
-      this.seil_1_4_5.setPos(0.0F, 0.0F, 7.0F);
-      this.seil_1_4_5.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 6.0F, 0.0F, 0.0F, 0.0F);
-      this.botom_3 = new ModelRenderer(this, 0, 0);
-      this.botom_3.setPos(18.0F, 3.0F, 17.0F);
-      this.botom_3.addBox(-14.0F, -9.0F, -3.0F, 28.0F, 16.0F, 3.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.botom_3, 1.5707964F, 0.0F, 0.0F);
-      this.Cargo1 = new ModelRenderer(this, 50, 47);
-      this.Cargo1.setPos(-0.8F, 0.0F, 15.8F);
-      this.Cargo1.addBox(0.0F, 0.0F, 0.0F, 7.0F, 3.0F, 13.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Cargo1, 0.0F, -1.5707964F, 0.0F);
-      this.Segel_1_3_1 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_3_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 9.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_3_1, -0.033161256F, 0.010471976F, 0.017453292F);
-      this.side_2 = new ModelRenderer(this, 0, 35);
-      this.side_2.setPos(0.0F, 4.0F, -9.0F);
-      this.side_2.addBox(-14.0F, -7.0F, -1.0F, 28.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_2, 0.0F, 3.1415927F, 0.0F);
-      this.botom_1_1 = new ModelRenderer(this, 0, 0);
-      this.botom_1_1.setPos(18.0F, 3.0F, 1.0F);
-      this.botom_1_1.addBox(-14.0F, -9.0F, -3.0F, 28.0F, 16.0F, 3.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.botom_1_1, 1.5707964F, 0.0F, 0.0F);
-      this.ruder_l_3 = new ModelRenderer(this, 62, 0);
-      this.ruder_l_3.setPos(34.0F, -4.0F, 25.0F);
-      this.ruder_l_3.addBox(-1.0F, 0.0F, -8.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_l_3.addBox(-0.8F, -3.0F, 9.0F, 1.0F, 6.0F, 7.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.ruder_l_3, -0.65449846F, 0.66095614F, 0.19634955F);
-      this.seil_1_4_4 = new ModelRenderer(this, 106, 0);
-      this.seil_1_4_4.setPos(0.0F, 0.0F, 11.0F);
-      this.seil_1_4_4.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 7.4F, 0.0F, 0.0F, 0.0F);
-      this.seil_4_4 = new ModelRenderer(this, 96, 0);
-      this.seil_4_4.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_4_4.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.seil_1_4_3 = new ModelRenderer(this, 96, 0);
-      this.seil_1_4_3.setPos(0.0F, 0.0F, 11.0F);
-      this.seil_1_4_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12 = new ModelRenderer(this, 97, 0);
-      this.Segel_1_12.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_12, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.botom_8 = new ModelRenderer(this, 10, 0);
-      this.botom_8.setPos(0.0F, 3.0F, 1.0F);
-      this.botom_8.addBox(-14.0F, -9.0F, -3.0F, 18.0F, 16.0F, 3.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.botom_8, 1.5707964F, 0.0F, 0.0F);
-      this.Cargo_0_1 = new ModelRenderer(this, 30, 55);
-      this.Cargo_0_1.setPos(12.3F, 3.0F, 1.0F);
-      this.Cargo_0_1.addBox(0.0F, 0.0F, 0.0F, 4.0F, 5.0F, 4.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_13_2.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, -2.0F, 11.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, -2.0F, 22.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, -2.0F, 33.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, -2.0F, 77.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, -2.0F, 66.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, -2.0F, 55.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_2.addBox(0.0F, -2.0F, 44.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_13_2, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.seil_2_3 = new ModelRenderer(this, 97, 0);
-      this.seil_2_3.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_2_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.seil_4_4_3 = new ModelRenderer(this, 111, 5);
-      this.seil_4_4_3.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_4_4_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 6.0F, 0.0F, 0.0F, 0.0F);
-      this.seil_2_4_2 = new ModelRenderer(this, 97, 0);
-      this.seil_2_4_2.setPos(0.0F, 0.0F, 13.0F);
-      this.seil_2_4_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9 = new ModelRenderer(this, 96, 0);
-      this.Segel_1_9.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_9, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_1_7_2 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_7_2.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7_2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_7_2, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.ruder_r_3 = new ModelRenderer(this, 62, 20);
-      this.ruder_r_3.setPos(34.0F, -4.0F, -10.0F);
-      this.ruder_r_3.addBox(-1.0F, 0.0F, -8.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_r_3.addBox(-0.2F, -3.0F, 9.0F, 1.0F, 6.0F, 7.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.ruder_r_3, -0.65449846F, 2.4806366F, 0.19634955F);
-      this.side_19 = new ModelRenderer(this, 0, 43);
-      this.side_19.setPos(56.0F, 4.0F, -9.0F);
-      this.side_19.addBox(-14.0F, -7.0F, -1.0F, 19.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_13_3.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, -2.0F, 11.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, -2.0F, 22.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, -2.0F, 33.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, -2.0F, 77.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, -2.0F, 66.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, -2.0F, 55.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_3.addBox(0.0F, -2.0F, 44.0F, 3.0F, 5.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_13_3, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.seil_4_3 = new ModelRenderer(this, 96, 0);
-      this.seil_4_3.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_4_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.side_10 = new ModelRenderer(this, 0, 27);
-      this.side_10.setPos(-17.0F, 4.0F, 8.0F);
-      this.side_10.addBox(-8.0F, -7.0F, -1.0F, 16.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_10, 0.0F, 1.5707964F, 0.0F);
-      this.Segel_1_2 = new ModelRenderer(this, 100, 0);
-      this.Segel_1_2.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 9.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_2, -0.033161256F, 0.010471976F, 0.017453292F);
-      this.Segel_1_4_3 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_4_3.setPos(0.0F, 0.0F, -2.0F);
-      this.Segel_1_4_3.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_3.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_4_3, -0.033161256F, 0.010471976F, 0.13962634F);
-      this.seil_2_4_3 = new ModelRenderer(this, 97, 0);
-      this.seil_2_4_3.setPos(0.0F, 0.0F, 13.0F);
-      this.seil_2_4_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 6.0F, 0.0F, 0.0F, 0.0F);
-      this.side_1 = new ModelRenderer(this, 0, 27);
-      this.side_1.setPos(63.0F, 4.0F, 8.0F);
-      this.side_1.addBox(-8.0F, -7.0F, -1.0F, 16.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_1, 0.0F, 1.5707964F, 0.0F);
-      this.Segel_1_16 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_16.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_16.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_16, -0.041887905F, 0.010471976F, 0.17453292F);
-      this.Segel_1_3_2 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_3_2.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 9.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_3_2, -0.033161256F, 0.010471976F, 0.017453292F);
-      this.Segel_1_12_1 = new ModelRenderer(this, 97, 0);
-      this.Segel_1_12_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_12_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_12_1, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_z3 = new ModelRenderer(this, 99, 0);
-      this.Segel_z3.setPos(2.0F, 3.0F, -60.0F);
-      this.Segel_z3.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, -2.0F, 76.0F, 1.0F, 4.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z3.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_z3, 0.0F, 0.0F, -0.57595867F);
-      this.botom_1 = new ModelRenderer(this, 10, -8);
-      this.botom_1.setPos(-8.0F, 15.0F, 28.0F);
-      this.botom_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.botom_1, 0.0F, 1.5707964F, 0.0F);
-      this.seil_2_4 = new ModelRenderer(this, 97, 0);
-      this.seil_2_4.setPos(0.0F, 0.0F, 13.0F);
-      this.seil_2_4.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.side_9 = new ModelRenderer(this, 0, 27);
-      this.side_9.setPos(-15.0F, 4.0F, 16.0F);
-      this.side_9.addBox(-8.0F, -7.0F, -1.0F, 16.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_9, 0.0F, 1.5707964F, 0.0F);
-      this.seil_1_4 = new ModelRenderer(this, 0, 0);
-      this.seil_1_4.setPos(0.0F, 0.0F, 0.0F);
-      this.seil_1_4.texOffs(100, 0).addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_l_1 = new ModelRenderer(this, 62, 0);
-      this.ruder_l_1.setPos(0.0F, -4.0F, 25.0F);
-      this.ruder_l_1.addBox(-1.0F, 0.0F, -8.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_l_1.addBox(-0.8F, -3.0F, 9.0F, 1.0F, 6.0F, 7.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.ruder_l_1, -0.65449846F, 0.66095614F, 0.19634955F);
-      this.mast_3_1 = new ModelRenderer(this, 10, 0);
-      this.mast_3_1.setPos(0.0F, -21.0F, 0.0F);
-      this.mast_3_1.addBox(0.0F, 0.0F, 0.0F, 3.0F, 21.0F, 4.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z0 = new ModelRenderer(this, 99, 0);
-      this.Segel_z0.setPos(1.5F, 2.0F, 0.0F);
-      this.Segel_z0.addBox(0.0F, 0.0F, -55.0F, 3.0F, 6.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z0.addBox(0.0F, 0.0F, -45.0F, 3.0F, 6.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z0.addBox(0.0F, 0.0F, -35.0F, 3.0F, 6.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z0.addBox(0.0F, 0.0F, -25.0F, 3.0F, 6.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z0.addBox(0.0F, 0.0F, 15.0F, 3.0F, 6.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z0.addBox(0.0F, 0.0F, 5.0F, 3.0F, 6.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z0.addBox(0.0F, 0.0F, -15.0F, 3.0F, 6.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z0.addBox(0.0F, 0.0F, -5.0F, 3.0F, 6.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.botom_7 = new ModelRenderer(this, 0, 0);
-      this.botom_7.setPos(29.0F, 6.0F, 9.0F);
-      this.botom_7.addBox(-14.0F, -9.0F, -3.0F, 28.0F, 16.0F, 3.0F, 0.0F, 4.0F, 0.0F);
-      this.setRotateAngle(this.botom_7, 1.5707964F, 0.0F, 0.0F);
-      this.Segel_1_3 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_3.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 9.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_3, -0.033161256F, 0.010471976F, 0.017453292F);
-      this.mast_1 = new ModelRenderer(this, 10, 0);
-      this.mast_1.setPos(18.0F, -31.0F, 6.0F);
-      this.mast_1.addBox(0.0F, 0.0F, 0.0F, 3.0F, 18.0F, 4.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3 = new ModelRenderer(this, 100, 0);
-      this.Segel_1_5_3.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_3.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_5_3, -0.033161256F, 0.010471976F, 0.08726646F);
-      this.side_5 = new ModelRenderer(this, 0, 43);
-      this.side_5.setPos(28.0F, 4.0F, 24.0F);
-      this.side_5.addBox(-14.0F, -7.0F, -1.0F, 28.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.side_4 = new ModelRenderer(this, 0, 35);
-      this.side_4.setPos(28.0F, 4.0F, -9.0F);
-      this.side_4.addBox(-14.0F, -7.0F, -1.0F, 28.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_4, 0.0F, 3.1415927F, 0.0F);
-      this.Cargo_1_2 = new ModelRenderer(this, 100, 25);
-      this.Cargo_1_2.mirror = true;
-      this.Cargo_1_2.setPos(0.5F, -5.0F, 7.0F);
-      this.Cargo_1_2.addBox(0.0F, 0.0F, -2.0F, 6.0F, 5.0F, 8.0F, 0.0F, 0.0F, -0.0F);
-      this.Segel_z1 = new ModelRenderer(this, 99, 0);
-      this.Segel_z1.setPos(2.0F, 3.0F, -60.0F);
-      this.Segel_z1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, -2.0F, 76.0F, 1.0F, 4.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_z1, 0.0F, 0.0F, -0.57595867F);
-      this.Segel_1_14 = new ModelRenderer(this, 97, 0);
-      this.Segel_1_14.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_14.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_14, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.side_3 = new ModelRenderer(this, 0, 43);
-      this.side_3.setPos(0.0F, 4.0F, 24.0F);
-      this.side_3.addBox(-14.0F, -7.0F, -1.0F, 28.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1 = new ModelRenderer(this, 96, 0);
-      this.Segel_1_9_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_9_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_9_1, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.mast_3 = new ModelRenderer(this, 10, 0);
-      this.mast_3.setPos(0.0F, -18.0F, 0.0F);
-      this.mast_3.addBox(0.0F, 0.0F, 0.0F, 3.0F, 18.0F, 4.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_11.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_11, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_1_2_2 = new ModelRenderer(this, 100, 0);
-      this.Segel_1_2_2.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 9.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_2_2, -0.033161256F, 0.010471976F, 0.017453292F);
-      this.side_11 = new ModelRenderer(this, 0, 27);
-      this.side_11.setPos(68.0F, -4.0F, 12.0F);
-      this.side_11.addBox(-8.0F, -7.0F, -1.0F, 16.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_11, 1.5707964F, 0.0F, 0.0F);
-      this.side_13 = new ModelRenderer(this, 0, 27);
-      this.side_13.setPos(64.9F, 1.0F, 12.0F);
-      this.side_13.addBox(-8.0F, -6.9F, -1.0F, 16.0F, 5.8F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.side_13, 1.5707964F, 0.0F, -0.6981317F);
-      this.Segel_z2 = new ModelRenderer(this, 99, 0);
-      this.Segel_z2.setPos(2.0F, 3.0F, -60.0F);
-      this.Segel_z2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, -2.0F, 76.0F, 1.0F, 4.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_z2, 0.0F, 0.0F, -0.57595867F);
-      this.Mast_unten = new ModelRenderer(this, 0, 5);
-      this.Mast_unten.setPos(21.0F, -25.0F, 8.5F);
-      this.Mast_unten.addBox(0.0F, 0.0F, -9.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_unten.addBox(0.0F, 0.0F, -27.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_unten.addBox(0.0F, 0.0F, 9.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_unten.addBox(0.0F, 0.0F, -45.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_unten.addBox(0.0F, 0.0F, -51.0F, 2.0F, 2.0F, 6.0F, 0.0F, 0.0F, 0.0F);
-      this.Mast_unten.addBox(0.0F, 0.0F, 27.0F, 2.0F, 2.0F, 12.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Mast_unten, 0.0F, -1.0471976F, 0.0F);
-      this.Cargo_0_2 = new ModelRenderer(this, 30, 55);
-      this.Cargo_0_2.setPos(8.0F, 3.0F, 0.0F);
-      this.Cargo_0_2.addBox(0.0F, 0.0F, 0.0F, 4.0F, 5.0F, 4.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_10.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_10, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.seil_1_2 = new ModelRenderer(this, 96, 0);
-      this.seil_1_2.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_1_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_8_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_8_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_8_1, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_1_6 = new ModelRenderer(this, 98, 0);
-      this.Segel_1_6.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_6, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.botom_2 = new ModelRenderer(this, 0, 0);
-      this.botom_2.setPos(46.0F, 3.0F, 1.0F);
-      this.botom_2.addBox(-14.0F, -9.0F, -3.0F, 28.0F, 16.0F, 3.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.botom_2, 1.5707964F, 0.0F, 0.0F);
-      this.ruder_l_4 = new ModelRenderer(this, 62, 0);
-      this.ruder_l_4.setPos(51.0F, -4.0F, 25.0F);
-      this.ruder_l_4.addBox(-1.0F, 0.0F, -8.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_l_4.addBox(-0.8F, -3.0F, 9.0F, 1.0F, 6.0F, 7.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.ruder_l_4, -0.65449846F, 0.66095614F, 0.19634955F);
-      this.Segel_1_5_2 = new ModelRenderer(this, 100, 0);
-      this.Segel_1_5_2.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_5_2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_5_2, -0.033161256F, 0.010471976F, 0.08726646F);
-      this.Segel_1_10_1 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_10_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_10_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_10_1, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_1_4_1 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_4_1.setPos(0.0F, 0.0F, -2.0F);
-      this.Segel_1_4_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_4_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_4_1, -0.033161256F, 0.010471976F, 0.13962634F);
-      this.botom_4 = new ModelRenderer(this, 0, 0);
-      this.botom_4.setPos(46.0F, 3.0F, 17.0F);
-      this.botom_4.addBox(-14.0F, -9.0F, -3.0F, 28.0F, 16.0F, 3.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.botom_4, 1.5707964F, 0.0F, 0.0F);
-      this.Segel_z4 = new ModelRenderer(this, 99, 0);
-      this.Segel_z4.setPos(2.0F, 3.0F, -60.0F);
-      this.Segel_z4.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, -2.0F, 76.0F, 1.0F, 4.0F, 10.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_z4.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_z4, 0.0F, 0.0F, -0.57595867F);
-      this.Segel_1_7 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_7.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_7.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_7, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_1_17 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_17.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_17.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_17, -0.041887905F, 0.010471976F, 0.17453292F);
-      this.Segel_1_6_2 = new ModelRenderer(this, 98, 0);
-      this.Segel_1_6_2.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_6_2.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_6_2, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.Segel_1_15 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_15.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_15.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_15, -0.041887905F, 0.010471976F, 0.17453292F);
-      this.Segel_1_13_1 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_13_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, -2.0F, 11.0F, 3.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, -2.0F, 22.0F, 3.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, -2.0F, 33.0F, 3.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, -2.0F, 77.0F, 3.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, -2.0F, 66.0F, 3.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, -2.0F, 55.0F, 3.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_13_1.addBox(0.0F, -2.0F, 44.0F, 3.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_13_1, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.seil_4_1 = new ModelRenderer(this, 96, 0);
-      this.seil_4_1.setPos(-31.0F, 28.5F, 17.5F);
-      this.seil_4_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.seil_4_1, 1.7872171F, 0.0F, 0.43458697F);
-      this.Segel_1_2_1 = new ModelRenderer(this, 100, 0);
-      this.Segel_1_2_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 9.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_2_1, -0.033161256F, 0.010471976F, 0.017453292F);
-      this.ruder_r_2 = new ModelRenderer(this, 62, 20);
-      this.ruder_r_2.setPos(17.0F, -4.0F, -10.0F);
-      this.ruder_r_2.addBox(-1.0F, 0.0F, -8.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_r_2.addBox(-0.2F, -3.0F, 9.0F, 1.0F, 6.0F, 7.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.ruder_r_2, -0.65449846F, 2.4806366F, 0.19634955F);
-      this.side_18 = new ModelRenderer(this, 0, 43);
-      this.side_18.setPos(56.0F, 4.0F, 24.0F);
-      this.side_18.addBox(-14.0F, -7.0F, -1.0F, 19.0F, 6.0F, 2.0F, 0.0F, 0.0F, 0.0F);
-      this.seil_1_4_1 = new ModelRenderer(this, 96, 0);
-      this.seil_1_4_1.setPos(0.0F, 0.0F, 0.0F);
-      this.seil_1_4_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_3_3.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 9.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_3_3.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_3_3, -0.033161256F, 0.010471976F, 0.017453292F);
-      this.Segel_1_2_3 = new ModelRenderer(this, 100, 0);
-      this.Segel_1_2_3.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 9.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_2_3.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_2_3, -0.033161256F, 0.010471976F, 0.017453292F);
-      this.Segel_1_11_1 = new ModelRenderer(this, 99, 0);
-      this.Segel_1_11_1.setPos(0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, -2.0F, 11.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, -2.0F, 0.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, -2.0F, 22.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, -2.0F, 33.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, -2.0F, 77.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, -2.0F, 66.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, -2.0F, 55.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.Segel_1_11_1.addBox(0.0F, -2.0F, 44.0F, 1.0F, 4.0F, 11.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.Segel_1_11_1, -0.041887905F, 0.010471976F, 0.08726646F);
-      this.botom_5 = new ModelRenderer(this, 0, 0);
-      this.botom_5.setPos(1.0F, 6.0F, 9.0F);
-      this.botom_5.addBox(-14.0F, -9.0F, -3.0F, 28.0F, 16.0F, 3.0F, 0.0F, 4.0F, 0.0F);
-      this.setRotateAngle(this.botom_5, 1.5707964F, 0.0F, 0.0F);
-      this.ruder_r_4 = new ModelRenderer(this, 62, 20);
-      this.ruder_r_4.setPos(51.0F, -4.0F, -10.0F);
-      this.ruder_r_4.addBox(-1.0F, 0.0F, -8.0F, 2.0F, 2.0F, 18.0F, 0.0F, 0.0F, 0.0F);
-      this.ruder_r_4.addBox(-0.2F, -3.0F, 9.0F, 1.0F, 6.0F, 7.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.ruder_r_4, -0.65449846F, 2.4806366F, 0.19634955F);
-      this.seil_4_2 = new ModelRenderer(this, 96, 0);
-      this.seil_4_2.setPos(0.0F, 0.0F, 14.0F);
-      this.seil_4_2.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.seil_1_1 = new ModelRenderer(this, 96, 0);
-      this.seil_1_1.setPos(-13.5F, -2.5F, -9.5F);
-      this.seil_1_1.addBox(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 14.0F, 0.0F, 0.0F, 0.0F);
-      this.setRotateAngle(this.seil_1_1, 1.3473941F, 0.0F, 0.43982297F);
-      this.botom_1.addChild(this.side_6);
-      this.botom_1.addChild(this.side_7);
-      this.seil_1_2.addChild(this.seil_1_3);
-      this.Segel_1_6_1.addChild(this.Segel_1_7_1);
-      this.Segel_1_7.addChild(this.Segel_1_8);
-      this.botom_1.addChild(this.seil_2_1);
-      this.Cargo1.addChild(this.Cargo_1_1);
-      this.Segel_1_4.addChild(this.Segel_1_5);
-      this.botom_1.addChild(this.Cargo0);
-      this.mast_3_1.addChild(this.mast_4);
-      this.Segel_1_3_2.addChild(this.Segel_1_4_2);
-      this.mast_1.addChild(this.mast_2);
-      this.Segel_1_4_1.addChild(this.Segel_1_5_1);
-      this.Segel_1_7_2.addChild(this.Segel_1_8_2);
-      this.Segel_1_17.addChild(this.Segel_1_18);
-      this.botom_1.addChild(this.side_8);
-      this.botom_1.addChild(this.botom_9);
-      this.seil_2_1.addChild(this.seil_2_2);
-      this.seil_1_4_1.addChild(this.seil_1_4_2);
-      this.Segel_1_3.addChild(this.Segel_1_4);
-      this.Segel_1_5_1.addChild(this.Segel_1_6_1);
-      this.seil_4_4.addChild(this.seil_4_4_2);
-      this.botom_1.addChild(this.ruder_l_2);
-      this.botom_1.addChild(this.ruder_r_1);
-      this.botom_1.addChild(this.Mast_oben);
-      this.Segel_1_12.addChild(this.Segel_1_13);
-      this.seil_2_4.addChild(this.seil_2_4_1);
-      this.botom_1.addChild(this.botom_6);
-      this.botom_1.addChild(this.side_12);
-      this.seil_1_4_4.addChild(this.seil_1_4_5);
-      this.botom_1.addChild(this.botom_3);
-      this.botom_1.addChild(this.Cargo1);
-      this.Segel_1_2_1.addChild(this.Segel_1_3_1);
-      this.botom_1.addChild(this.side_2);
-      this.botom_1.addChild(this.botom_1_1);
-      this.botom_1.addChild(this.ruder_l_3);
-      this.seil_1_4_3.addChild(this.seil_1_4_4);
-      this.seil_4_3.addChild(this.seil_4_4);
-      this.seil_1_4_2.addChild(this.seil_1_4_3);
-      this.Segel_1_11.addChild(this.Segel_1_12);
-      this.botom_1.addChild(this.botom_8);
-      this.Cargo0.addChild(this.Cargo_0_1);
-      this.Segel_1_8_2.addChild(this.Segel_1_13_2);
-      this.seil_2_2.addChild(this.seil_2_3);
-      this.seil_4_4_2.addChild(this.seil_4_4_3);
-      this.seil_2_4_1.addChild(this.seil_2_4_2);
-      this.Segel_1_8.addChild(this.Segel_1_9);
-      this.Segel_1_6_2.addChild(this.Segel_1_7_2);
-      this.botom_1.addChild(this.ruder_r_3);
-      this.botom_1.addChild(this.side_19);
-      this.Segel_1_5_3.addChild(this.Segel_1_13_3);
-      this.seil_4_2.addChild(this.seil_4_3);
-      this.botom_1.addChild(this.side_10);
-      this.Segel_z4.addChild(this.Segel_1_2);
-      this.Segel_1_3_3.addChild(this.Segel_1_4_3);
-      this.seil_2_4_2.addChild(this.seil_2_4_3);
-      this.botom_1.addChild(this.side_1);
-      this.Segel_1_15.addChild(this.Segel_1_16);
-      this.Segel_1_2_2.addChild(this.Segel_1_3_2);
-      this.Segel_1_11_1.addChild(this.Segel_1_12_1);
-      this.Mast_oben.addChild(this.Segel_z3);
-      this.seil_2_3.addChild(this.seil_2_4);
-      this.botom_1.addChild(this.side_9);
-      this.seil_1_3.addChild(this.seil_1_4);
-      this.botom_1.addChild(this.ruder_l_1);
-      this.mast_3.addChild(this.mast_3_1);
-      this.Mast_oben.addChild(this.Segel_z0);
-      this.botom_1.addChild(this.botom_7);
-      this.Segel_1_2.addChild(this.Segel_1_3);
-      this.botom_1.addChild(this.mast_1);
-      this.Segel_1_4_3.addChild(this.Segel_1_5_3);
-      this.botom_1.addChild(this.side_5);
-      this.botom_1.addChild(this.side_4);
-      this.Cargo1.addChild(this.Cargo_1_2);
-      this.Mast_oben.addChild(this.Segel_z1);
-      this.Segel_1_13.addChild(this.Segel_1_14);
-      this.botom_1.addChild(this.side_3);
-      this.Segel_1_8_1.addChild(this.Segel_1_9_1);
-      this.mast_2.addChild(this.mast_3);
-      this.Segel_1_10.addChild(this.Segel_1_11);
-      this.Segel_z2.addChild(this.Segel_1_2_2);
-      this.botom_1.addChild(this.side_11);
-      this.botom_1.addChild(this.side_13);
-      this.Mast_oben.addChild(this.Segel_z2);
-      this.botom_1.addChild(this.Mast_unten);
-      this.Cargo0.addChild(this.Cargo_0_2);
-      this.Segel_1_9.addChild(this.Segel_1_10);
-      this.seil_1_1.addChild(this.seil_1_2);
-      this.Segel_1_7_1.addChild(this.Segel_1_8_1);
-      this.Segel_1_5.addChild(this.Segel_1_6);
-      this.botom_1.addChild(this.botom_2);
-      this.botom_1.addChild(this.ruder_l_4);
-      this.Segel_1_4_2.addChild(this.Segel_1_5_2);
-      this.Segel_1_9_1.addChild(this.Segel_1_10_1);
-      this.Segel_1_3_1.addChild(this.Segel_1_4_1);
-      this.botom_1.addChild(this.botom_4);
-      this.Mast_oben.addChild(this.Segel_z4);
-      this.Segel_1_6.addChild(this.Segel_1_7);
-      this.Segel_1_16.addChild(this.Segel_1_17);
-      this.Segel_1_5_2.addChild(this.Segel_1_6_2);
-      this.Segel_1_14.addChild(this.Segel_1_15);
-      this.Segel_1_12_1.addChild(this.Segel_1_13_1);
-      this.mast_1.addChild(this.seil_4_1);
-      this.Segel_z3.addChild(this.Segel_1_2_1);
-      this.botom_1.addChild(this.ruder_r_2);
-      this.botom_1.addChild(this.side_18);
-      this.seil_1_4.addChild(this.seil_1_4_1);
-      this.Segel_1_2_3.addChild(this.Segel_1_3_3);
-      this.Segel_z1.addChild(this.Segel_1_2_3);
-      this.Segel_1_10_1.addChild(this.Segel_1_11_1);
-      this.botom_1.addChild(this.botom_5);
-      this.botom_1.addChild(this.ruder_r_4);
-      this.seil_4_1.addChild(this.seil_4_2);
-      this.botom_1.addChild(this.seil_1_1);
+        this.root = new ModelRenderer(this);
+        this.root.setPos(0.0F, 24.0F, 0.0F);
 
-      this.rightPaddles = new ModelRenderer[]{this.ruder_r_1, this.ruder_r_2, this.ruder_r_3, this.ruder_r_4};
-      this.leftPaddles = new ModelRenderer[]{this.ruder_l_1, this.ruder_l_2, this.ruder_l_3, this.ruder_l_4};
-      this.sailStates = new ModelRenderer[]{this.Segel_z0, this.Segel_z1, this.Segel_z2, this.Segel_z3, this.Segel_z4};
-   }
+        this.galley = new ModelRenderer(this);
+        this.galley.setPos(0.0F, 0.0F, 0.0F);
+        this.root.addChild(this.galley);
 
-   @Override
-   public void renderToBuffer(MatrixStack matrixStackIn, IVertexBuilder bufferIn, int packedLightIn, int packedOverlayIn, float red, float green, float blue, float alpha) {
-      boolean[] visibility = captureVisibility(this.sailStates);
+        ModelRenderer deck = new ModelRenderer(this);
+        deck.setPos(14.0F, 0.0F, 0.0F);
+        this.galley.addChild(deck);
 
-      setVisibility(this.sailStates, false);
+        ModelRenderer cube_r1 = new ModelRenderer(this);
+        cube_r1.setPos(-14.0F, 0.0F, 0.0F);
+        deck.addChild(cube_r1);
+        setRotationAngle(cube_r1, 1.5708F, 0.0F, 0.0F);
+        cube_r1.texOffs(28, 0).addBox(-42.0F, -13.0F, 2.0F, 14.0F, 13.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(28, 0).addBox(-42.0F, 0.0F, 2.0F, 14.0F, 13.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(33, 0).addBox(-53.0F, 0.0F, 4.0F, 8.0F, 6.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(33, 0).addBox(-53.0F, -6.0F, 4.0F, 8.0F, 6.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(33, 0).addBox(-64.0F, -3.5F, 4.0F, 11.0F, 6.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(15, 0).addBox(38.0F, -13.0F, 2.0F, 13.0F, 13.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(15, 0).addBox(38.0F, 0.0F, 2.0F, 13.0F, 13.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(18, 0).addBox(28.0F, -16.0F, 2.0F, 10.0F, 16.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(18, 0).addBox(28.0F, 0.0F, 2.0F, 10.0F, 16.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(0, 0).addBox(0.0F, 0.0F, 2.0F, 28.0F, 16.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(0, 0).addBox(0.0F, -16.0F, 2.0F, 28.0F, 16.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(0, 0).addBox(-28.0F, 0.0F, 2.0F, 28.0F, 16.0F, 3.0F, 0.0F, false);
+        cube_r1.texOffs(0, 0).addBox(-28.0F, -16.0F, 2.0F, 28.0F, 16.0F, 3.0F, 0.0F, false);
 
-      ImmutableList.of(this.botom_1).forEach((modelRenderer) -> {
-         modelRenderer.render(matrixStackIn, bufferIn, packedLightIn, packedOverlayIn, red, green, blue, alpha);
-      });
+        ModelRenderer bottom = new ModelRenderer(this);
+        bottom.setPos(0.0F, 0.0F, 0.0F);
+        this.galley.addChild(bottom);
 
-      restoreVisibility(this.sailStates, visibility);
-   }
+        ModelRenderer cube_r2 = new ModelRenderer(this);
+        cube_r2.setPos(0.0F, 0.0F, 0.0F);
+        bottom.addChild(cube_r2);
+        cube_r2.texOffs(8, 2).addBox(-42.0F, 3.0F, -4.0F, 14.0F, 7.0F, 8.0F, 0.0F, false);
 
-   @Override
-   public void setupAnim(AbstractBannerUser entityIn, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
-      if (!(entityIn instanceof GalleyEntity)) {
-         return;
-      }
+        ModelRenderer cube_r5_r1 = new ModelRenderer(this);
+        cube_r5_r1.setPos(-42.0F, 5.5F, 0.0F);
+        cube_r2.addChild(cube_r5_r1);
+        setRotationAngle(cube_r5_r1, 0.0F, 0.0F, 1.5708F);
+        cube_r5_r1.texOffs(4, 1).addBox(-11.6F, 0.8F, -3.0F, 10.0F, 7.0F, 6.0F, 0.0F, false);
 
-      GalleyEntity galley = (GalleyEntity) entityIn;
+        ModelRenderer cube_r4_r1 = new ModelRenderer(this);
+        cube_r4_r1.setPos(-42.0F, 5.5F, 0.0F);
+        cube_r2.addChild(cube_r4_r1);
+        setRotationAngle(cube_r4_r1, 0.0F, 0.0F, 0.6545F);
+        cube_r4_r1.texOffs(4, 1).addBox(-7.2F, -3.5F, -3.5F, 10.0F, 7.0F, 7.0F, 0.0F, false);
 
-      this.botom_1.xRot = 0.0F;
-      this.botom_1.yRot = HULL_Y_ROT;
-      this.botom_1.zRot = 0.0F;
+        ModelRenderer cube_r3 = new ModelRenderer(this);
+        cube_r3.setPos(0.0F, 0.0F, 0.0F);
+        bottom.addChild(cube_r3);
+        setRotationAngle(cube_r3, 1.5708F, 0.0F, 0.0F);
+        cube_r3.texOffs(0, 0).addBox(32.0F, -4.0F, -10.0F, 15.0F, 8.0F, 7.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(16.0F, -4.0F, -10.0F, 16.0F, 8.0F, 7.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(-6.0F, -4.0F, -10.0F, 22.0F, 8.0F, 7.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(-28.0F, -4.0F, -10.0F, 22.0F, 8.0F, 7.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(-28.0F, 2.0F, -3.0F, 22.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(-28.0F, -12.0F, -3.0F, 22.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(-42.0F, -12.0F, -3.0F, 14.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(-42.0F, 2.0F, -3.0F, 14.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(-6.0F, -12.0F, -3.0F, 24.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(34.0F, -12.0F, -3.0F, 17.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(-6.0F, 2.0F, -3.0F, 24.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(34.0F, 2.0F, -3.0F, 17.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(18.0F, 2.0F, -3.0F, 16.0F, 10.0F, 5.0F, 0.0F, false);
+        cube_r3.texOffs(0, 0).addBox(18.0F, -12.0F, -3.0F, 16.0F, 10.0F, 5.0F, 0.0F, false);
 
-      int state = galley.getSailState();
-      updateSailState(state, this.sailStates);
+        ModelRenderer sides = new ModelRenderer(this);
+        sides.setPos(0.0F, 0.0F, 0.0F);
+        this.galley.addChild(sides);
+        sides.texOffs(8, 36).addBox(-43.0F, -11.0F, -16.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(-43.0F, -11.0F, 13.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(37.0F, -11.0F, -16.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(37.0F, -17.0F, -16.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(13.0F, -11.0F, -19.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(-2.0F, -11.0F, -19.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(-28.0F, -11.0F, -19.0F, 11.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(-17.0F, -11.0F, -19.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(37.0F, -11.0F, 13.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(28.0F, -11.0F, 16.0F, 9.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(28.0F, -11.0F, -19.0F, 9.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(37.0F, -17.0F, 13.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(13.0F, -11.0F, 16.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(-2.0F, -11.0F, 16.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(-17.0F, -11.0F, 16.0F, 15.0F, 6.0F, 3.0F, 0.0F, false);
+        sides.texOffs(8, 36).addBox(-28.0F, -11.0F, 16.0F, 11.0F, 6.0F, 3.0F, 0.0F, false);
 
-      int cargo = galley.getCargo();
-      this.Cargo0.visible = cargo >= 1;
-      this.Cargo1.visible = cargo >= 2;
-      animatePaddles(galley, 0, limbSwing, this.leftPaddles, false);
-      animatePaddles(galley, 1, limbSwing, this.rightPaddles, true);
-   }
+        ModelRenderer cube_r4 = new ModelRenderer(this);
+        cube_r4.setPos(35.0F, -8.0F, -9.5F);
+        sides.addChild(cube_r4);
+        setRotationAngle(cube_r4, 0.0F, 1.5708F, 0.0F);
 
-   public void setRotateAngle(ModelRenderer modelRenderer, float x, float y, float z) {
-      modelRenderer.xRot = x;
-      modelRenderer.yRot = y;
-      modelRenderer.zRot = z;
-   }
+        ModelRenderer cube_r5 = new ModelRenderer(this);
+        cube_r5.setPos(-20.5F, -8.0F, -3.5F);
+        sides.addChild(cube_r5);
+        setRotationAngle(cube_r5, 0.0F, -1.5708F, 0.0F);
+        cube_r5.texOffs(8, 36).addBox(3.5001F, -3.0F, 21.5F, 13.0F, 6.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(-9.4999F, -3.0F, 21.5F, 13.0F, 6.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(3.5001F, 3.0F, 21.5F, 13.0F, 3.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(-9.4999F, 3.0F, 21.5F, 13.0F, 3.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(-8.4999F, 6.0F, 21.5F, 11.0F, 3.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(3.5001F, 6.0F, 21.5F, 12.0F, 3.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(-9.5002F, -3.0F, -74.5F, 13.0F, 6.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(-9.5002F, -9.0F, -74.5F, 13.0F, 6.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(3.4998F, -3.0F, -74.5F, 13.0F, 6.0F, 3.0F, 0.0F, false);
+        cube_r5.texOffs(8, 36).addBox(3.4998F, -9.0F, -74.5F, 13.0F, 6.0F, 3.0F, 0.0F, false);
 
-   private void animatePaddles(GalleyEntity galleyEntity, int side, float limbSwing, ModelRenderer[] paddles, boolean rightSide) {
-      float rowingTime = galleyEntity.getRowingTime(side, limbSwing);
-      float steer = -galleyEntity.getRotSpeed();
+        ModelRenderer mast_1 = new ModelRenderer(this);
+        mast_1.setPos(14.0F, -5.0F, -1.0F);
+        this.galley.addChild(mast_1);
+        mast_1.texOffs(8, 0).addBox(-13.0F, -15.0F, -0.5F, 3.0F, 15.0F, 3.0F, 0.0F, false);
+        mast_1.texOffs(8, 0).addBox(-13.0F, -30.0F, -0.5F, 3.0F, 15.0F, 3.0F, 0.0F, false);
+        mast_1.texOffs(8, 0).addBox(-13.0F, -45.0F, -0.5F, 3.0F, 15.0F, 3.0F, 0.0F, false);
+        mast_1.texOffs(8, 0).addBox(-13.0F, -60.0F, -0.5F, 3.0F, 15.0F, 3.0F, 0.0F, false);
+        mast_1.texOffs(8, 0).addBox(-13.0F, -75.0F, -0.5F, 3.0F, 15.0F, 3.0F, 0.0F, false);
 
-      for (ModelRenderer paddle : paddles) {
-         setPaddleRotation(paddle, rowingTime, steer, rightSide);
-      }
-   }
+        ModelRenderer mast_2 = new ModelRenderer(this);
+        mast_2.setPos(-11.0F, -65.5F, 3.4F);
+        mast_1.addChild(mast_2);
+        setRotationAngle(mast_2, 0.0F, 0.0F, 0.7418F);
+        mast_2.texOffs(8, 0).addBox(1.2957F, -39.0539F, -0.9F, 2.0F, 15.0F, 4.0F, 0.0F, false);
+        mast_2.texOffs(8, 0).addBox(1.2957F, -52.0539F, -0.9F, 2.0F, 13.0F, 4.0F, 0.0F, false);
+        mast_2.texOffs(8, 0).addBox(1.2957F, -24.0539F, -0.9F, 2.0F, 15.0F, 4.0F, 0.0F, false);
+        mast_2.texOffs(8, 0).addBox(1.2957F, -9.0539F, -0.9F, 2.0F, 15.0F, 4.0F, 0.0F, false);
+        mast_2.texOffs(8, 0).addBox(1.2957F, 5.9461F, -0.9F, 2.0F, 15.0F, 4.0F, 0.0F, false);
+        mast_2.texOffs(8, 0).addBox(1.2957F, 20.9461F, -0.9F, 2.0F, 15.0F, 4.0F, 0.0F, false);
+        mast_2.texOffs(8, 0).addBox(1.2957F, 35.9461F, -0.9F, 2.0F, 15.0F, 4.0F, 0.0F, false);
 
-   private static void setPaddleRotation(ModelRenderer paddle, float rowingTime, float steer, boolean rightSide) {
-      float xRot = (float)MathHelper.clamp(PADDLE_MIN_X_ROT, PADDLE_MAX_X_ROT, (double)((MathHelper.sin(-rowingTime) + 1.0F) / 2.0F));
-      float yawOffset = (float)MathHelper.clamp(PADDLE_MIN_Y_ROT, PADDLE_MAX_Y_ROT, (double)((MathHelper.sin(-rowingTime + 1.0F) + 1.0F) / 2.0F));
-      if (rightSide) {
-         paddle.xRot = xRot;
-         paddle.yRot = (float)Math.PI - yawOffset + steer;
-      } else {
-         paddle.xRot = xRot;
-         paddle.yRot = yawOffset - steer;
-      }
-   }
+        ModelRenderer cube_r6 = new ModelRenderer(this);
+        cube_r6.setPos(29.0F, 0.5F, 1.0F);
+        mast_1.addChild(cube_r6);
+        setRotationAngle(cube_r6, 0.0F, 0.0F, 0.5672F);
+        cube_r6.texOffs(8, 0).addBox(6.6905F, -13.6357F, -2.5F, 4.0F, 24.0F, 5.0F, 0.0F, false);
+        cube_r6.texOffs(8, 0).addBox(7.4905F, -23.3357F, -1.0F, 2.0F, 10.0F, 2.0F, 0.0F, false);
 
-   private static boolean[] captureVisibility(ModelRenderer[] sails) {
-      boolean[] visibility = new boolean[sails.length];
-      for (int i = 0; i < sails.length; ++i) {
-         visibility[i] = sails[i].visible;
-      }
-      return visibility;
-   }
+        ModelRenderer bannerStick = new ModelRenderer(this);
+        bannerStick.setPos(-14.0F, 4.0F, 1.0F);
+        mast_1.addChild(bannerStick);
+        bannerStick.texOffs(8, 0).addBox(2.0F, -94.0F, -0.5F, 1.0F, 15.0F, 1.0F, 0.0F, false);
 
-   private static void setVisibility(ModelRenderer[] sails, boolean visible) {
-      for (ModelRenderer sail : sails) {
-         sail.visible = visible;
-      }
-   }
+        ModelRenderer chest_1 = new ModelRenderer(this);
+        chest_1.setPos(1.0F, 0.0F, 1.0F);
+        this.galley.addChild(chest_1);
+        chest_1.texOffs(96, 38).addBox(33.0F, -13.0F, -13.0F, 8.0F, 8.0F, 8.0F, 0.0F, false);
 
-   private static void restoreVisibility(ModelRenderer[] sails, boolean[] visibility) {
-      for (int i = 0; i < sails.length; ++i) {
-         sails[i].visible = visibility[i];
-      }
-   }
+        ModelRenderer cube_r7 = new ModelRenderer(this);
+        cube_r7.setPos(27.75F, -11.75F, -8.25F);
+        chest_1.addChild(cube_r7);
+        setRotationAngle(cube_r7, 0.0F, 1.5708F, 0.0F);
+        cube_r7.texOffs(30, 55).addBox(-2.75F, -4.25F, 6.75F, 4.0F, 3.0F, 4.0F, 0.0F, false);
 
-   private static void updateSailState(int state, ModelRenderer[] sails) {
-      int clamped = MathHelper.clamp(state, 0, sails.length - 1);
-      for (int i = 0; i < sails.length; ++i) {
-         sails[i].visible = i == clamped;
-      }
-   }
+        ModelRenderer cube_r8 = new ModelRenderer(this);
+        cube_r8.setPos(19.0F, -15.5F, -18.0F);
+        chest_1.addChild(cube_r8);
+        setRotationAngle(cube_r8, 1.5708F, -1.5708F, 0.0F);
+        cube_r8.texOffs(30, 55).addBox(15.0001F, -24.5001F, -10.0001F, 4.0F, 5.0F, 4.0F, 0.0F, false);
+
+        ModelRenderer cube_r9 = new ModelRenderer(this);
+        cube_r9.setPos(31.0F, 9.0F, -75.0F);
+        chest_1.addChild(cube_r9);
+        setRotationAngle(cube_r9, 0.0F, -1.5708F, 0.0F);
+        cube_r9.texOffs(30, 55).addBox(72.1001F, -19.0F, -17.4001F, 4.0F, 5.0F, 4.0F, 0.0F, false);
+
+        ModelRenderer cube_r10 = new ModelRenderer(this);
+        cube_r10.setPos(22.0F, -9.0F, -1.0F);
+        chest_1.addChild(cube_r10);
+        setRotationAngle(cube_r10, 0.0F, 3.1416F, 0.0F);
+        cube_r10.texOffs(96, 38).addBox(-28.0F, -4.0F, 3.5F, 8.0F, 8.0F, 8.0F, 0.0F, false);
+
+        ModelRenderer chest_4 = new ModelRenderer(this);
+        chest_4.setPos(42.0F, -11.5F, 9.0F);
+        this.galley.addChild(chest_4);
+
+        ModelRenderer cube_r11 = new ModelRenderer(this);
+        cube_r11.setPos(0.0F, 0.0F, 0.0F);
+        chest_4.addChild(cube_r11);
+        setRotationAngle(cube_r11, 0.0F, 1.5708F, 0.0F);
+        cube_r11.texOffs(64, 29).addBox(-3.0F, -1.5F, -4.0F, 6.0F, 5.0F, 8.0F, 0.0F, false);
+
+        ModelRenderer cube_r12 = new ModelRenderer(this);
+        cube_r12.setPos(3.0F, 20.5F, -42.0F);
+        chest_4.addChild(cube_r12);
+        setRotationAngle(cube_r12, 0.0F, -1.5708F, 0.0F);
+        cube_r12.texOffs(30, 55).addBox(39.0F, -22.0F, -5.0F, 4.0F, 5.0F, 4.0F, 0.0F, false);
+
+        ModelRenderer cube_r13 = new ModelRenderer(this);
+        cube_r13.setPos(0.0F, 20.5F, -42.0F);
+        chest_4.addChild(cube_r13);
+        setRotationAngle(cube_r13, 0.0F, -1.5708F, 0.0F);
+        cube_r13.texOffs(50, 47).addBox(38.0F, -17.0F, -9.25F, 7.0F, 3.0F, 13.0F, 0.0F, false);
+
+        ModelRenderer chest_2 = new ModelRenderer(this);
+        chest_2.setPos(-55.0F, -15.5F, 30.0F);
+        this.galley.addChild(chest_2);
+
+        ModelRenderer cube_r14 = new ModelRenderer(this);
+        cube_r14.setPos(0.0F, 0.0F, 0.0F);
+        chest_2.addChild(cube_r14);
+        setRotationAngle(cube_r14, 1.5708F, -1.5708F, 0.0F);
+        cube_r14.texOffs(30, 55).addBox(-21.0001F, -24.4999F, -9.9999F, 4.0F, 5.0F, 4.0F, 0.0F, false);
+
+        ModelRenderer cube_r15 = new ModelRenderer(this);
+        cube_r15.setPos(12.0F, 24.5F, -57.0F);
+        chest_2.addChild(cube_r15);
+        setRotationAngle(cube_r15, 0.0F, -1.5708F, 0.0F);
+        cube_r15.texOffs(30, 55).addBox(30.9999F, -19.0F, -14.9999F, 4.0F, 5.0F, 4.0F, 0.0F, false);
+        cube_r15.texOffs(30, 55).addBox(32.0F, -25.0F, -8.0F, 4.0F, 3.0F, 4.0F, 0.0F, false);
+
+        ModelRenderer cube_r16 = new ModelRenderer(this);
+        cube_r16.setPos(3.0F, 6.5F, -16.0F);
+        chest_2.addChild(cube_r16);
+        setRotationAngle(cube_r16, 0.0F, 3.1416F, 0.0F);
+        cube_r16.texOffs(96, 38).addBox(-19.0F, -4.0F, 3.5F, 8.0F, 8.0F, 8.0F, 0.0F, false);
+
+        ModelRenderer chest_3 = new ModelRenderer(this);
+        chest_3.setPos(37.3333F, -8.0F, -11.1667F);
+        this.galley.addChild(chest_3);
+
+        ModelRenderer cube_r17 = new ModelRenderer(this);
+        cube_r17.setPos(-20.3333F, 17.0F, -33.8333F);
+        chest_3.addChild(cube_r17);
+        setRotationAngle(cube_r17, 0.0F, -1.5708F, 0.0F);
+        cube_r17.texOffs(30, 55).addBox(33.0F, -19.0F, 53.0F, 4.0F, 5.0F, 4.0F, 0.0F, false);
+        cube_r17.texOffs(31, 56).addBox(33.0F, -19.0F, 49.0F, 4.0F, 5.0F, 3.0F, 0.0F, false);
+
+        ModelRenderer cube_r18 = new ModelRenderer(this);
+        cube_r18.setPos(-29.3333F, -1.0F, 7.1667F);
+        chest_3.addChild(cube_r18);
+        setRotationAngle(cube_r18, 0.0F, 1.5708F, 0.0F);
+        cube_r18.texOffs(96, 38).addBox(-4.0F, -4.0F, -49.0F, 8.0F, 8.0F, 8.0F, 0.0F, false);
+
+        this.steer = new ModelRenderer(this);
+        this.steer.setPos(38.0F, 3.25F, 0.0F);
+        this.galley.addChild(this.steer);
+        this.steer.texOffs(4, 1).addBox(13.0F, 2.75F, -1.0F, 4.0F, 8.0F, 2.0F, 0.0F, false);
+        this.steer.texOffs(4, 1).addBox(9.0F, -0.25F, -1.0F, 4.0F, 11.0F, 2.0F, 0.0F, false);
+
+        ModelRenderer row_L_4 = new ModelRenderer(this);
+        row_L_4.setPos(-12.5F, -12.0F, -18.0F);
+        this.galley.addChild(row_L_4);
+        setRotationAngle(row_L_4, 0.2618F, 0.0F, 0.0F);
+        row_L_4.texOffs(33, 3).addBox(-2.5F, -0.5F, -28.0F, 5.0F, 1.0F, 10.0F, 0.0F, false);
+
+        ModelRenderer cube_r19 = new ModelRenderer(this);
+        cube_r19.setPos(-4.5F, 9.0F, -23.0F);
+        row_L_4.addChild(cube_r19);
+        setRotationAngle(cube_r19, 1.5708F, 0.0F, 0.0F);
+        cube_r19.texOffs(9, 0).addBox(4.0F, 18.5F, 8.0F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+        cube_r19.texOffs(9, 0).addBox(4.0F, 1.5F, 8.0F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+
+        ModelRenderer row_L_3 = new ModelRenderer(this);
+        row_L_3.setPos(1.5F, -12.0F, -18.0F);
+        this.galley.addChild(row_L_3);
+        setRotationAngle(row_L_3, 0.2618F, 0.0F, 0.0F);
+        row_L_3.texOffs(33, 3).addBox(-2.5F, -0.5F, -28.0F, 5.0F, 1.0F, 10.0F, 0.0F, false);
+
+        ModelRenderer cube_r20 = new ModelRenderer(this);
+        cube_r20.setPos(-4.5F, 9.0F, -23.0F);
+        row_L_3.addChild(cube_r20);
+        setRotationAngle(cube_r20, 1.5708F, 0.0F, 0.0F);
+        cube_r20.texOffs(9, 0).addBox(4.0F, 18.5F, 8.0F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+        cube_r20.texOffs(9, 0).addBox(4.0F, 1.5F, 8.0F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+
+        ModelRenderer row_L_2 = new ModelRenderer(this);
+        row_L_2.setPos(16.5F, -12.0F, -18.0F);
+        this.galley.addChild(row_L_2);
+        setRotationAngle(row_L_2, 0.2618F, 0.0F, 0.0F);
+        row_L_2.texOffs(33, 3).addBox(-2.5F, -0.5F, -28.0F, 5.0F, 1.0F, 10.0F, 0.0F, false);
+
+        ModelRenderer cube_r21 = new ModelRenderer(this);
+        cube_r21.setPos(-4.5F, 9.0F, -23.0F);
+        row_L_2.addChild(cube_r21);
+        setRotationAngle(cube_r21, 1.5708F, 0.0F, 0.0F);
+        cube_r21.texOffs(9, 0).addBox(4.0F, 19.0F, 8.0F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+        cube_r21.texOffs(9, 0).addBox(4.0F, 2.0F, 8.0F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+
+        ModelRenderer row_L_1 = new ModelRenderer(this);
+        row_L_1.setPos(31.5F, -12.0F, -18.0F);
+        this.galley.addChild(row_L_1);
+        setRotationAngle(row_L_1, 0.2618F, 0.0F, 0.0F);
+        row_L_1.texOffs(33, 3).addBox(-2.5F, -0.5F, -28.0F, 5.0F, 1.0F, 10.0F, 0.0F, false);
+
+        ModelRenderer cube_r22 = new ModelRenderer(this);
+        cube_r22.setPos(-4.5F, 9.0F, -23.0F);
+        row_L_1.addChild(cube_r22);
+        setRotationAngle(cube_r22, 1.5708F, 0.0F, 0.0F);
+        cube_r22.texOffs(9, 0).addBox(4.0F, 18.5F, 8.0F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+        cube_r22.texOffs(9, 0).addBox(4.0F, 1.5F, 8.0F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+
+        ModelRenderer row_R_1 = new ModelRenderer(this);
+        row_R_1.setPos(31.5F, -11.0F, 18.0F);
+        this.galley.addChild(row_R_1);
+        setRotationAngle(row_R_1, -0.2618F, 0.0F, 0.0F);
+        row_R_1.texOffs(33, 3).addBox(-2.5F, -0.2059F, 11.1704F, 5.0F, 1.0F, 10.0F, 0.0F, false);
+
+        ModelRenderer cube_r23 = new ModelRenderer(this);
+        cube_r23.setPos(-2.5F, 8.0F, -16.0F);
+        row_R_1.addChild(cube_r23);
+        setRotationAngle(cube_r23, 1.5708F, 0.0F, 0.0F);
+        cube_r23.texOffs(9, 0).addBox(2.0F, 13.6704F, 6.7059F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+        cube_r23.texOffs(9, 0).addBox(2.0F, -3.3296F, 6.7059F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+
+        ModelRenderer row_R_2 = new ModelRenderer(this);
+        row_R_2.setPos(16.5F, -11.0F, 18.0F);
+        this.galley.addChild(row_R_2);
+        setRotationAngle(row_R_2, -0.2618F, 0.0F, 0.0F);
+        row_R_2.texOffs(33, 3).addBox(-2.5F, -0.2059F, 11.1704F, 5.0F, 1.0F, 10.0F, 0.0F, false);
+
+        ModelRenderer cube_r24 = new ModelRenderer(this);
+        cube_r24.setPos(-4.5F, 8.0F, -16.0F);
+        row_R_2.addChild(cube_r24);
+        setRotationAngle(cube_r24, 1.5708F, 0.0F, 0.0F);
+        cube_r24.texOffs(9, 0).addBox(4.0F, 13.6704F, 6.7059F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+        cube_r24.texOffs(9, 0).addBox(4.0F, -3.3296F, 6.7059F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+
+        ModelRenderer row_R_3 = new ModelRenderer(this);
+        row_R_3.setPos(1.5F, -11.0F, 18.0F);
+        this.galley.addChild(row_R_3);
+        setRotationAngle(row_R_3, -0.2618F, 0.0F, 0.0F);
+        row_R_3.texOffs(33, 3).addBox(-2.5F, -0.2059F, 11.1704F, 5.0F, 1.0F, 10.0F, 0.0F, false);
+
+        ModelRenderer cube_r25 = new ModelRenderer(this);
+        cube_r25.setPos(-4.5F, 8.0F, -16.0F);
+        row_R_3.addChild(cube_r25);
+        setRotationAngle(cube_r25, 1.5708F, 0.0F, 0.0F);
+        cube_r25.texOffs(9, 0).addBox(4.0F, 13.6704F, 6.7059F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+        cube_r25.texOffs(9, 0).addBox(4.0F, -3.3296F, 6.7059F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+
+        ModelRenderer row_R_4 = new ModelRenderer(this);
+        row_R_4.setPos(-12.5F, -11.0F, 18.0F);
+        this.galley.addChild(row_R_4);
+        setRotationAngle(row_R_4, -0.2618F, 0.0F, 0.0F);
+        row_R_4.texOffs(33, 3).addBox(-2.5F, -0.2059F, 11.1704F, 5.0F, 1.0F, 10.0F, 0.0F, false);
+
+        ModelRenderer cube_r26 = new ModelRenderer(this);
+        cube_r26.setPos(-4.5F, 8.0F, -16.0F);
+        row_R_4.addChild(cube_r26);
+        setRotationAngle(cube_r26, 1.5708F, 0.0F, 0.0F);
+        cube_r26.texOffs(9, 0).addBox(4.0F, 13.6704F, 6.7059F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+        cube_r26.texOffs(9, 0).addBox(4.0F, -3.3296F, 6.7059F, 1.0F, 17.0F, 2.0F, 0.0F, false);
+
+        this.chests = new ModelRenderer[]{chest_1, chest_2, chest_3, chest_4};
+        this.leftPaddles = new ModelRenderer[]{row_L_1, row_L_2, row_L_3, row_L_4};
+        this.rightPaddles = new ModelRenderer[]{row_R_1, row_R_2, row_R_3, row_R_4};
+    }
+
+    @Override
+    public void renderToBuffer(MatrixStack matrixStack, IVertexBuilder buffer, int packedLight, int packedOverlay, float red, float green, float blue, float alpha) {
+        this.root.render(matrixStack, buffer, packedLight, packedOverlay, red, green, blue, alpha);
+    }
+
+    @Override
+    public void setupAnim(AbstractBannerUser entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
+        if (!(entity instanceof GalleyEntity)) {
+            return;
+        }
+
+        GalleyEntity galleyEntity = (GalleyEntity) entity;
+        this.root.xRot = 0.0F;
+        this.root.yRot = HULL_Y_ROT;
+        this.root.zRot = 0.0F;
+
+        int cargo = galleyEntity.getCargo();
+        for (int i = 0; i < this.chests.length; ++i) {
+            this.chests[i].visible = cargo > i;
+        }
+
+        animatePaddles(galleyEntity, 0, limbSwing, this.leftPaddles, false);
+        animatePaddles(galleyEntity, 1, limbSwing, this.rightPaddles, true);
+    }
+
+    private void animatePaddles(GalleyEntity galleyEntity, int side, float limbSwing, ModelRenderer[] paddles, boolean rightSide) {
+        float rowingTime = galleyEntity.getRowingTime(side, limbSwing);
+        float steerValue = -galleyEntity.getRotSpeed();
+        for (ModelRenderer paddle : paddles) {
+            setPaddleRotation(paddle, rowingTime, steerValue, rightSide);
+        }
+    }
+
+    private static void setPaddleRotation(ModelRenderer paddle, float rowingTime, float steer, boolean rightSide) {
+        float xRot = MathHelper.clamp((MathHelper.sin(-rowingTime) + 1.0F) / 2.0F, PADDLE_MIN_X_ROT, PADDLE_MAX_X_ROT);
+        float yawOffset = MathHelper.clamp((MathHelper.sin(-rowingTime + 1.0F) + 1.0F) / 2.0F, PADDLE_MIN_Y_ROT, PADDLE_MAX_Y_ROT);
+        if (rightSide) {
+            paddle.xRot = xRot;
+            paddle.yRot = (float) Math.PI - yawOffset + steer;
+        } else {
+            paddle.xRot = xRot;
+            paddle.yRot = yawOffset - steer;
+        }
+    }
+
+    private static void setRotationAngle(ModelRenderer modelRenderer, float x, float y, float z) {
+        modelRenderer.xRot = x;
+        modelRenderer.yRot = y;
+        modelRenderer.zRot = z;
+    }
 }
